+++ src/context/core.rs
@@ -1,13 +1,12 @@
 use super::{
     Align, BasicCompositeOperation, Command, CompositeOperation, CompositeOperationState, FillType,
-    ImageFlags, ImageId, LineCap, LineJoin, Paint, PathDir, TextureType, KAPPA90,
+    LineCap, LineJoin, Paint, PathDir,
 };
 use crate::fonts::{FontId, Fonts, LayoutChar};
 use crate::renderer::Scissor;
 use crate::Color;
 use crate::{cache::PathCache, Extent, Point, Rect, Renderer, Transform};
 use clamped::Clamp;
-use std::f32::consts::PI;
 
 #[derive(Clone)]
 pub(super) struct State {
@@ -237,61 +236,6 @@ impl<R: Renderer> Context<R> {
         self.state_mut().fill_type = fill_type;
     }
 
-    pub fn create_image<D: AsRef<[u8]>>(
-        &mut self,
-        flags: ImageFlags,
-        data: D,
-    ) -> anyhow::Result<ImageId> {
-        let img = image::load_from_memory(data.as_ref())?;
-        let img = img.to_rgba();
-        let dimensions = img.dimensions();
-        let img = self.renderer.create_texture(
-            TextureType::RGBA,
-            dimensions.0 as usize,
-            dimensions.1 as usize,
-            flags,
-            Some(&img.into_raw()),
-        )?;
-        Ok(img)
-    }
-
-    pub fn create_image_from_file<P: AsRef<std::path::Path>>(
-        &mut self,
-        flags: ImageFlags,
-        path: P,
-    ) -> anyhow::Result<ImageId> {
-        self.create_image(flags, std::fs::read(path)?)
-    }
-
-    pub fn create_image_rgba(
-        &mut self,
-        width: usize,
-        height: usize,
-        flags: ImageFlags,
-        data: Option<&[u8]>,
-    ) -> anyhow::Result<ImageId> {
-        let img = self
-            .renderer
-            .create_texture(TextureType::RGBA, width, height, flags, data)?;
-        Ok(img)
-    }
-
-    pub fn update_image(&mut self, img: ImageId, data: &[u8]) -> anyhow::Result<()> {
-        let (w, h) = self.renderer.texture_size(img.clone())?;
-        self.renderer.update_texture(img, 0, 0, w, h, data)?;
-        Ok(())
-    }
-
-    pub fn image_size(&self, img: ImageId) -> anyhow::Result<(usize, usize)> {
-        let res = self.renderer.texture_size(img)?;
-        Ok(res)
-    }
-
-    pub fn delete_image(&mut self, img: ImageId) -> anyhow::Result<()> {
-        self.renderer.delete_texture(img)?;
-        Ok(())
-    }
-
     pub fn scissor<T: Into<Rect>>(&mut self, rect: T) {
         let rect = rect.into();
         let state = self.state_mut();
@@ -344,321 +288,8 @@ impl<R: Renderer> Context<R> {
         self.state_mut().composite_operation = op.into();
     }
 
-    fn append_command(&mut self, cmd: Command) {
-        let state = self.states.last().unwrap();
-        let xform = &state.xform;
-        match cmd {
-            Command::MoveTo(pt) => {
-                self.commands
-                    .push(Command::MoveTo(xform.transform_point(pt)));
-                self.last_position = pt;
-            }
-            Command::LineTo(pt) => {
-                self.commands
-                    .push(Command::LineTo(xform.transform_point(pt)));
-                self.last_position = pt;
-            }
-            Command::BezierTo(pt1, pt2, pt3) => {
-                self.last_position = pt3;
-                self.commands.push(Command::BezierTo(
-                    xform.transform_point(pt1),
-                    xform.transform_point(pt2),
-                    xform.transform_point(pt3),
-                ));
-            }
-            _ => {
-                self.commands.push(cmd);
-            }
-        }
-    }
-
-    pub fn begin_path(&mut self) {
-        self.commands.clear();
-        self.cache.clear();
-    }
-
-    pub fn move_to<P: Into<Point>>(&mut self, pt: P) {
-        self.append_command(Command::MoveTo(pt.into()));
-    }
-
-    pub fn line_to<P: Into<Point>>(&mut self, pt: P) {
-        self.append_command(Command::LineTo(pt.into()));
-    }
-
-    pub fn bezier_to<P: Into<Point>>(&mut self, cp1: P, cp2: P, pt: P) {
-        self.append_command(Command::BezierTo(cp1.into(), cp2.into(), pt.into()));
-    }
-
-    pub fn quad_to<P: Into<Point>>(&mut self, cp: P, pt: P) {
-        let x0 = self.last_position.x;
-        let y0 = self.last_position.y;
-        let cp = cp.into();
-        let pt = pt.into();
-        self.append_command(Command::BezierTo(
-            Point::new(x0 + 2.0 / 3.0 * (cp.x - x0), y0 + 2.0 / 3.0 * (cp.y - y0)),
-            Point::new(
-                pt.x + 2.0 / 3.0 * (cp.x - pt.x),
-                pt.y + 2.0 / 3.0 * (cp.y - pt.y),
-            ),
-            pt,
-        ));
-    }
-
-    pub fn arc_to<P: Into<Point>>(&mut self, pt1: P, pt2: P, radius: f32) {
-        let pt0 = self.last_position;
-
-        if self.commands.is_empty() {
-            return;
-        }
-
-        let pt1 = pt1.into();
-        let pt2 = pt2.into();
-        if pt0.equals(pt1, self.dist_tol)
-            || pt1.equals(pt2, self.dist_tol)
-            || pt1.dist_pt_seg(pt0, pt2) < self.dist_tol * self.dist_tol
-            || radius < self.dist_tol
-        {
-            self.line_to(pt1);
-            return;
-        }
-
-        let d0 = Point::new(pt0.x - pt1.x, pt0.y - pt1.y);
-        let d1 = Point::new(pt2.x - pt1.x, pt2.y - pt1.y);
-        let a = (d0.x * d1.x + d0.y * d1.y).cos();
-        let d = radius / (a / 2.0).tan();
-
-        if d > 10000.0 {
-            self.line_to(pt1);
-            return;
-        }
-
-        let (cx, cy, a0, a1, dir) = if Point::cross(d0, d1) > 0.0 {
-            (
-                pt1.x + d0.x * d + d0.y * radius,
-                pt1.y + d0.y * d + -d0.x * radius,
-                d0.x.atan2(-d0.y),
-                -d1.x.atan2(d1.y),
-                PathDir::CW,
-            )
-        } else {
-            (
-                pt1.x + d0.x * d + -d0.y * radius,
-                pt1.y + d0.y * d + d0.x * radius,
-                -d0.x.atan2(d0.y),
-                d1.x.atan2(-d1.y),
-                PathDir::CCW,
-            )
-        };
-
-        self.arc(Point::new(cx, cy), radius, a0, a1, dir);
-    }
-
-    pub fn close_path(&mut self) {
-        self.commands.push(Command::Close);
-    }
-
     pub fn path_winding<D: Into<PathDir>>(&mut self, dir: D) {
         self.commands.push(Command::Winding(dir.into()));
-    }
-
-    pub fn arc<P: Into<Point>>(&mut self, cp: P, radius: f32, a0: f32, a1: f32, dir: PathDir) {
-        let cp = cp.into();
-        let move_ = self.commands.is_empty();
-
-        let mut da = a1 - a0;
-        if dir == PathDir::CW {
-            if da.abs() >= PI * 2.0 {
-                da = PI * 2.0;
-            } else {
-                while da < 0.0 {
-                    da += PI * 2.0;
-                }
-            }
-        } else {
-            if da.abs() >= PI * 2.0 {
-                da = -PI * 2.0;
-            } else {
-                while da > 0.0 {
-                    da -= PI * 2.0;
-                }
-            }
-        }
-
-        let ndivs = ((da.abs() / (PI * 0.5) + 0.5) as i32).min(5).max(1);
-        let hda = (da / (ndivs as f32)) / 2.0;
-        let mut kappa = (4.0 / 3.0 * (1.0 - hda.cos()) / hda.sin()).abs();
-
-        if dir == PathDir::CCW {
-            kappa = -kappa;
-        }
-
-        let mut px = 0.0;
-        let mut py = 0.0;
-        let mut ptanx = 0.0;
-        let mut ptany = 0.0;
-
-        for i in 0..=ndivs {
-            let a = a0 + da * ((i as f32) / (ndivs as f32));
-            let dx = a.cos();
-            let dy = a.sin();
-            let x = cp.x + dx * radius;
-            let y = cp.y + dy * radius;
-            let tanx = -dy * radius * kappa;
-            let tany = dx * radius * kappa;
-
-            if i == 0 {
-                if move_ {
-                    self.append_command(Command::MoveTo(Point::new(x, y)));
-                } else {
-                    self.append_command(Command::LineTo(Point::new(x, y)));
-                }
-            } else {
-                self.append_command(Command::BezierTo(
-                    Point::new(px + ptanx, py + ptany),
-                    Point::new(x - tanx, y - tany),
-                    Point::new(x, y),
-                ));
-            }
-            px = x;
-            py = y;
-            ptanx = tanx;
-            ptany = tany;
-        }
-    }
-
-    pub fn rect<T: Into<Rect>>(&mut self, rect: T) {
-        let rect = rect.into();
-        self.append_command(Command::MoveTo(Point::new(rect.xy.x, rect.xy.y)));
-        self.append_command(Command::LineTo(Point::new(
-            rect.xy.x,
-            rect.xy.y + rect.size.height,
-        )));
-        self.append_command(Command::LineTo(Point::new(
-            rect.xy.x + rect.size.width,
-            rect.xy.y + rect.size.height,
-        )));
-        self.append_command(Command::LineTo(Point::new(
-            rect.xy.x + rect.size.width,
-            rect.xy.y,
-        )));
-        self.append_command(Command::Close);
-    }
-
-    pub fn rounded_rect<T: Into<Rect>>(&mut self, rect: T, radius: f32) {
-        let rect = rect.into();
-        self.rounded_rect_varying(rect, radius, radius, radius, radius);
-    }
-
-    pub fn rounded_rect_varying<T: Into<Rect>>(
-        &mut self,
-        rect: T,
-        lt: f32,
-        rt: f32,
-        rb: f32,
-        lb: f32,
-    ) {
-        let rect = rect.into();
-        if lt < 0.1 && rt < 0.1 && lb < 0.1 && rb < 0.1 {
-            self.rect(rect);
-        } else {
-            let halfw = rect.size.width.abs() * 0.5;
-            let halfh = rect.size.height.abs() * 0.5;
-            let rxlb = lb.min(halfw) * rect.size.width.signum();
-            let rylb = lb.min(halfh) * rect.size.height.signum();
-            let rxrb = rb.min(halfw) * rect.size.width.signum();
-            let ryrb = rb.min(halfh) * rect.size.height.signum();
-            let rxrt = rt.min(halfw) * rect.size.width.signum();
-            let ryrt = rt.min(halfh) * rect.size.height.signum();
-            let rxlt = lt.min(halfw) * rect.size.width.signum();
-            let rylt = lt.min(halfh) * rect.size.height.signum();
-
-            self.append_command(Command::MoveTo(Point::new(rect.xy.x, rect.xy.y + rylt)));
-            self.append_command(Command::LineTo(Point::new(
-                rect.xy.x,
-                rect.xy.y + rect.size.height - rylb,
-            )));
-            self.append_command(Command::BezierTo(
-                Point::new(
-                    rect.xy.x,
-                    rect.xy.y + rect.size.height - rylb * (1.0 - KAPPA90),
-                ),
-                Point::new(
-                    rect.xy.x + rxlb * (1.0 - KAPPA90),
-                    rect.xy.y + rect.size.height,
-                ),
-                Point::new(rect.xy.x + rxlb, rect.xy.y + rect.size.height),
-            ));
-            self.append_command(Command::LineTo(Point::new(
-                rect.xy.x + rect.size.width - rxrb,
-                rect.xy.y + rect.size.height,
-            )));
-            self.append_command(Command::BezierTo(
-                Point::new(
-                    rect.xy.x + rect.size.width - rxrb * (1.0 - KAPPA90),
-                    rect.xy.y + rect.size.height,
-                ),
-                Point::new(
-                    rect.xy.x + rect.size.width,
-                    rect.xy.y + rect.size.height - ryrb * (1.0 - KAPPA90),
-                ),
-                Point::new(
-                    rect.xy.x + rect.size.width,
-                    rect.xy.y + rect.size.height - ryrb,
-                ),
-            ));
-            self.append_command(Command::LineTo(Point::new(
-                rect.xy.x + rect.size.width,
-                rect.xy.y + ryrt,
-            )));
-            self.append_command(Command::BezierTo(
-                Point::new(
-                    rect.xy.x + rect.size.width,
-                    rect.xy.y + ryrt * (1.0 - KAPPA90),
-                ),
-                Point::new(
-                    rect.xy.x + rect.size.width - rxrt * (1.0 - KAPPA90),
-                    rect.xy.y,
-                ),
-                Point::new(rect.xy.x + rect.size.width - rxrt, rect.xy.y),
-            ));
-            self.append_command(Command::LineTo(Point::new(rect.xy.x + rxlt, rect.xy.y)));
-            self.append_command(Command::BezierTo(
-                Point::new(rect.xy.x + rxlt * (1.0 - KAPPA90), rect.xy.y),
-                Point::new(rect.xy.x, rect.xy.y + rylt * (1.0 - KAPPA90)),
-                Point::new(rect.xy.x, rect.xy.y + rylt),
-            ));
-            self.append_command(Command::Close);
-        }
-    }
-
-    pub fn ellipse<P: Into<Point>>(&mut self, center: P, radius_x: f32, radius_y: f32) {
-        let center = center.into();
-        self.append_command(Command::MoveTo(Point::new(center.x - radius_x, center.y)));
-        self.append_command(Command::BezierTo(
-            Point::new(center.x - radius_x, center.y + radius_y * KAPPA90),
-            Point::new(center.x - radius_x * KAPPA90, center.y + radius_y),
-            Point::new(center.x, center.y + radius_y),
-        ));
-        self.append_command(Command::BezierTo(
-            Point::new(center.x + radius_x * KAPPA90, center.y + radius_y),
-            Point::new(center.x + radius_x, center.y + radius_y * KAPPA90),
-            Point::new(center.x + radius_x, center.y),
-        ));
-        self.append_command(Command::BezierTo(
-            Point::new(center.x + radius_x, center.y - radius_y * KAPPA90),
-            Point::new(center.x + radius_x * KAPPA90, center.y - radius_y),
-            Point::new(center.x, center.y - radius_y),
-        ));
-        self.append_command(Command::BezierTo(
-            Point::new(center.x - radius_x * KAPPA90, center.y - radius_y),
-            Point::new(center.x - radius_x, center.y - radius_y * KAPPA90),
-            Point::new(center.x - radius_x, center.y),
-        ));
-        self.append_command(Command::Close);
-    }
-
-    pub fn circle<P: Into<Point>>(&mut self, center: P, radius: f32) {
-        self.ellipse(center.into(), radius, radius);
     }
 
     pub fn fill(&mut self) -> anyhow::Result<()> {
